+++ CHANGELOG.rst
@@ -2,10 +2,9 @@
 Changelog
 =========
 
-x.x.x (TBD)
-==================
 * Added ``hide`` parameter to ElasticsearchTarget
 * Added ExpressionTarget support for ElasticSearch data sources
+* Added ``legendCalcs`` parameter to TimeSeries Panel
 
 0.6.2 (2022-02-24)
 ==================